+++ scripts/gb_model/__init__.py
@@ -1,5 +1,3 @@
 # SPDX-FileCopyrightText: gb-open-market-model contributors
 #
 # SPDX-License-Identifier: MIT
-
-